fix(core-components-tabs): fix mobile\desktop split
