+++ packages/tabs/package.json
@@ -4,7 +4,7 @@
     "description": "Tabs components",
     "keywords": [],
     "license": "ISC",
-    "main": "dist/index.js",
+    "main": "dist/desktop.js",
     "types": "dist/typings.d.ts",
     "files": [
         "dist"
