+++ packages/tabs/src/Component.stories.mdx
@@ -2,16 +2,9 @@ import { Meta, Story, Props, Preview } from '@storybook/addon-docs/blocks';
 import { boolean, select } from '@storybook/addon-knobs';
 import { ComponentHeader } from 'storybook/blocks/component-header';
 
-import {
-    TabsDesktop,
-    TabsMobile,
-    TabsResponsive,
-    PrimaryTabListDesktop,
-    PrimaryTabListMobile,
-    SecondaryTabListDesktop,
-    SecondaryTabListMobile,
-    Tab,
-} from './index';
+import { Tab, TabsDesktop, PrimaryTabListDesktop, SecondaryTabListDesktop } from './desktop';
+import { TabsMobile, PrimaryTabListMobile, SecondaryTabListMobile } from './mobile';
+import { TabsResponsive } from './responsive';
 
 import { name, version } from '../package.json';
