--- packages/tabs/src/components/index.ts
@@ -1,5 +1,0 @@
-export * from './primary-tablist';
-export * from './secondary-tablist';
-export * from './scrollable-container';
-export * from './tab';
-export * from './tabs';
