+++ packages/tabs/src/components/secondary-tablist/Component.responsive.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { SecondaryTabListDesktop, SecondaryTabListMobile } from './index';
+import { SecondaryTabListDesktop } from './Component.desktop';
+import { SecondaryTabListMobile } from './Component.mobile';
 import { SecondaryTabListProps } from '../../typings';
 import { useWindowWidth } from '../../utils';
