+++ packages/tabs/src/components/tabs/Component.mobile.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TabsProps } from '../../typings';
-import { PrimaryTabListMobile } from '../primary-tablist';
-import { SecondaryTabListMobile } from '../secondary-tablist';
+import { PrimaryTabListMobile } from '../primary-tablist/Component.mobile';
+import { SecondaryTabListMobile } from '../secondary-tablist/Component.mobile';
 import { Tabs } from './Component';
 
 const views = {
