+++ packages/tabs/src/components/tabs/Component.responsive.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TabsProps } from '../../typings';
-import { PrimaryTabListResponsive } from '../primary-tablist';
-import { SecondaryTabListResponsive } from '../secondary-tablist';
+import { PrimaryTabListResponsive } from '../primary-tablist/Component.responsive';
+import { SecondaryTabListResponsive } from '../secondary-tablist/Component.responsive';
 import { Tabs } from './Component';
 
 const views = {
