--- packages/tabs/src/components/tabs/index.ts
@@ -1,4 +1,0 @@
-export * from './Component';
-export * from './Component.desktop';
-export * from './Component.mobile';
-export * from './Component.responsive';
