+++ packages/tabs/src/desktop.ts
@@ -1,0 +1,6 @@
+export * from './components/primary-tablist/Component.desktop';
+export * from './components/scrollable-container';
+export * from './components/secondary-tablist/Component.desktop';
+export * from './components/tab';
+export * from './components/tabs/Component.desktop';
+export * from './useTabs';
