+++ packages/tabs/src/mobile.ts
@@ -1,0 +1,6 @@
+export * from './components/primary-tablist/Component.mobile';
+export * from './components/scrollable-container';
+export * from './components/secondary-tablist/Component.mobile';
+export * from './components/tab';
+export * from './components/tabs/Component.mobile';
+export * from './useTabs';
