+++ packages/tabs/src/responsive.ts
@@ -1,0 +1,6 @@
+export * from './components/primary-tablist/Component.responsive';
+export * from './components/scrollable-container';
+export * from './components/secondary-tablist/Component.responsive';
+export * from './components/tab';
+export * from './components/tabs/Component.responsive';
+export * from './useTabs';
